requirements.txt
Flask
flask-cors
gunicorn
pytesseract
pdfplumber
Pillow
python-docx
